+++ components/StageView/StageMemberView.tsx
@@ -29,8 +29,8 @@ const StageMemberTitle = (props: { stageMember: StageMemberWithUserData; withIco
         justifyContent: 'center',
       }}
     >
-      {withIcon && <HiUserCircle size={80} sx={{ color: 'gray.3' }} />}
-      <Box sx={{ position: !withIcon ? 'absolute' : 'static', bottom: '0px', left: '50px' }}>
+      {withIcon && <HiUserCircle size={60} sx={{ color: 'gray.3' }} />}
+      <Box sx={{ position: !withIcon ? 'absolute' : 'static', bottom: '0px', left: '0px' }}>
         <Heading as="h5" sx={{ display: withIcon ? 'block' : 'inline-block' }}>
           <OnlineStatus online={stageMember.online} /> {stageMember.name}
         </Heading>
@@ -53,8 +53,10 @@ const StageMemberTitle = (props: { stageMember: StageMemberWithUserData; withIco
 
 const StageMemberView = ({
   stageMember,
+  variant,
 }: {
   stageMember: StageMemberWithUserData;
+  variant?: 'default' | 'conductor';
 }): JSX.Element => {
   const videoConsumers = useVideoConsumersByStageMember(stageMember._id);
 
@@ -63,8 +65,9 @@ const StageMemberView = ({
       sx={{
         position: 'relative',
         backgroundImage: videoConsumers.length <= 0 && 'url("/images/user_background.svg")',
-        width: '100%',
-        height: '100%',
+        height: variant === 'conductor' ? '100%' : '240px',
+        width: variant === 'conductor' ? '100%' : '240px',
+        m: 1,
       }}
     >
       {videoConsumers.length > 0 && (