+++ components/StageView/index.tsx
@@ -5,21 +5,21 @@ import React from 'react';
 import { Box, Flex, jsx } from 'theme-ui';
 import GroupView from './GroupView';
 import { useCurrentStageId, useGroupsByStage } from '../../lib/use-digital-stage/hooks';
-import ConductorsView from '../new/elements/StageView/ConductorsView';
+import ConductorsView from './ConductorsView';
 
-const StageViewTest = (): JSX.Element => {
+const StageView = (): JSX.Element => {
   const stageId = useCurrentStageId();
   const groups = useGroupsByStage(stageId);
 
-  console.log(groups);
+  console.log(groups.length);
 
   return (
     <Box
       sx={{
-        width: 'calc(100vw - [10px, 80px])',
+        width: ['calc(100vw - 10px)', 'calc(100vw - 80px)'],
         height: 'calc(100vh - 100px)',
-        py: '50px',
-        px: '50px',
+        pb: '50px',
+        px: '25px',
         overflowY: 'auto',
         '::-webkit-scrollbar': {
           width: '15px',
@@ -35,7 +35,7 @@ const StageViewTest = (): JSX.Element => {
         },
       }}
     >
-      <Flex sx={{ height: '100%', width: '100%', flexWrap: 'wrap' }}>
+      <Flex sx={{ flexWrap: 'wrap' }}>
         {groups &&
           groups.map((group) => {
             return <GroupView key={group._id} group={group} />;
@@ -46,4 +46,4 @@ const StageViewTest = (): JSX.Element => {
   );
 };
 
-export default StageViewTest;
+export default StageView;