+++ components/new/elements/Menu/AppBar/index.tsx
@@ -1,16 +1,16 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
-import * as React from 'react';
-import { jsx, Box, Button } from 'theme-ui';
+import { Fragment, useEffect, useRef, useState } from 'react';
 import { FaUser } from 'react-icons/fa';
-import DropdownMenu from './DropdownMenu';
+import { Box, Button, jsx } from 'theme-ui';
 import SettingsModal from '../../../../settings';
+import DropdownMenu from './DropdownMenu';
 
 const AppBar = (): JSX.Element => {
-  const [open, setOpen] = React.useState<boolean>(false);
-  const [openSettings, setOpenSettings] = React.useState<boolean>(false);
-  const [selected, setSelected] = React.useState<string>('');
-  const node = React.useRef(null);
+  const [open, setOpen] = useState<boolean>(false);
+  const [openSettings, setOpenSettings] = useState<boolean>(false);
+  const [selected, setSelected] = useState<string>('');
+  const node = useRef(null);
 
   const handleClick = (e) => {
     if (node && node.current.contains(e.target)) {
@@ -20,15 +20,27 @@ const AppBar = (): JSX.Element => {
     setOpen(false);
   };
 
-  React.useEffect(() => {
-    document.addEventListener('mousedown', handleClick);
-    return () => {
-      document.removeEventListener('mousedown', handleClick);
-    };
-  }, []);
+  const handleSelect = (selected) => {
+    if (selected !== 'feedback') {
+      setSelected(selected);
+      setOpenSettings(true);
+      setOpen(false);
+    } else if (selected === 'feedback') {
+      window.open('https://forum.digital-stage.org/c/deutsch/ds-web/30', '_target');
+    }
+  };
+
+  useEffect(() => {
+    if (open) {
+      document.addEventListener('mousedown', handleClick);
+      return () => {
+        document.removeEventListener('mousedown', handleClick);
+      };
+    }
+  }, [open]);
 
   return (
-    <React.Fragment>
+    <Fragment>
       <Box role="menu">
         <Button
           variant="circle"
@@ -38,7 +50,7 @@ const AppBar = (): JSX.Element => {
             bg: 'text',
             color: 'gray.2',
           }}
-          onClick={() => setOpen((prev) => !prev)}
+          onClick={() => setOpen(!open)}
         >
           <FaUser />
         </Button>
@@ -47,15 +59,7 @@ const AppBar = (): JSX.Element => {
         <DropdownMenu
           isOpen={open}
           onClose={() => setOpen(false)}
-          onSelect={(selected) => {
-            if (selected !== 'feedback') {
-              setSelected(selected);
-              setOpenSettings(true);
-              setOpen(false);
-            } else if (selected === 'feedback') {
-              window.open('https://forum.digital-stage.org/', '_target');
-            }
-          }}
+          onSelect={(selected) => handleSelect(selected)}
         />
       </div>
       <SettingsModal
@@ -63,7 +67,7 @@ const AppBar = (): JSX.Element => {
         onClose={() => setOpenSettings(false)}
         selected={selected}
       />
-    </React.Fragment>
+    </Fragment>
   );
 };
 export default AppBar;